refactor: getUsersData 이동
