+++ src/entities/user/index.ts
@@ -1,1 +1,2 @@
 export type { User } from "./model/types";
+export { getUsersData } from "./api/users.api";
